+++ apps/web/next.config.mjs
@@ -2,6 +2,10 @@ import { createVanillaExtractPlugin } from '@vanilla-extract/next-plugin';
 const withVanillaExtract = createVanillaExtractPlugin();
 
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  experimental: {
+    scrollRestoration: true,
+  },
+};
 
 export default withVanillaExtract(nextConfig);
