+++ packages/theme/src/index.ts
@@ -1,3 +1,3 @@
 export * from './tokens';
 export * from './themes';
-export * from './sprinkles.css';
+export * from './sprinkles';
