+++ packages/theme/src/sprinkles/index.ts
@@ -1,0 +1,1 @@
+export { sprinkles } from './sprinkles.css';
